% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_ligand_model.R
\name{formatForBLM}
\alias{formatForBLM}
\title{Format data for BLM model}
\usage{
formatForBLM(
  location,
  sample,
  analyte,
  result,
  ecoregion = "Willamette Valley",
  limit = TRUE,
  total = FALSE
)
}
\arguments{
\item{location}{a vector of location identifiers}

\item{sample}{a vector of sample identifiers}

\item{analyte}{a vector of analyte names: must match name variable of unexported \code{kBLMParams} data.frame}

\item{result}{a vector of concentrations}

\item{ecoregion}{the ecoregion to allow regional defaults. Currently not used.}

\item{limit}{should the output be limited to samples with dissolved copper (or total copper if \code{total = TRUE}).}

\item{total}{should missing dissolved copper by filled with total copper if available}
}
\description{
Format data for BLM model.
}
