% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotic_ligand_model.R
\name{imputeBLM}
\alias{imputeBLM}
\title{Fill missing values in BLM input data}
\usage{
imputeBLM(x, HA = 10, S = 1e-06)
}
\arguments{
\item{x}{a data.frame of BLM input data created by \link{formatForBLM}}

\item{HA}{a default value for humic acid}

\item{S}{a default value for sulfide ion}
}
\description{
Fill missing values in BLM input data using simple linear
regression on conductivity. These equations were fit by Oregon
Department of Environnmental Quality using data collected in Oregon.
}
