% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalCriteria.R
\name{metalCriteria}
\alias{metalCriteria}
\title{Calculate metals criteria for a vector for hardness concentrations and
analyte names}
\usage{
metalCriteria(hardness, analyte, toxicity = c("acute", "chronic"))
}
\arguments{
\item{hardness}{A vector of hardness values}

\item{analyte}{A vector of metal analyte names}

\item{toxicity}{whether to use acute or chronic toxicity}
}
\value{
A vector of hardness dependent metals criteria
}
\description{
Calculate metals criteria for a vector for hardness concentrations and
analyte names
}
\author{
Jason Law <jason.law@portlandoregon.gov>
}
