% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh4Criteria.R
\name{nh4Criteria}
\alias{nh4Criteria}
\title{Calculate Oregon DEQ water quality criteria for ammonia}
\usage{
nh4Criteria(pH, Cel, toxicity = c("chronic", "acute"), salmonids = FALSE)
}
\arguments{
\item{pH}{a water pH for the criteria}

\item{Cel}{a water temperature in degrees Celsius}

\item{toxicity}{the type of criteria, \code{toxicity = 'acute'} 
gives the acute criteria (CMC) while \code{toxicity = 'chronic'} gives the chronic criteria (CCC)}

\item{salmonids}{logical, whether the criteria should be for streams with salmonid 
beneficial use (Table 30A). The default is \code{FALSE}. Ignored when \code{toxicity = 'chronic'}.}
}
\value{
a numeric vector of water quality criteria
}
\description{
This function calculates the Oregon DEQ pH and temperature dependent
ammonia criteria. Tables of values and formulas can be found at 
\url{http://www.deq.state.or.us/wq/standards/docs/tables303140.pdf}.
}
\details{
Table 30A acute, salmonids are a designated use.
Table 30b acute, salmonids are not a designated use.
Table 30c chronic, fish and aquatic life are a designated use.
}
