\name{ORDEQWaterQualityCriteria-package}
\alias{ORDEQWaterQualityCriteria-package}
\alias{ORDEQWaterQualityCriteria}
\docType{package}
\title{
Functions for calculating Oregon Department of Environmental Quality Water
Quality Criteria.
}
\description{
Functions for calculating Oregon Department of Environmental Quality Water
Quality Criteria.  Mostly for hardness dependent metal criteria, but might
include more in the future.
}
\details{
\tabular{ll}{
Package: \tab ORDEQWaterQualityCriteria\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-07-01\cr
License: \tab GPL v3\cr
LazyLoad: \tab yes\cr
}
See add.metal.criteria to create a metals standards vector for inclusion with 
long format chemistry data.
}
\author{
Jason Law <jason.law@portlandoregon.gov
}
\references{
See Oregon DEQ web site: \url{http://www.deq.state.or.us/wq/standards/toxics.htm}
}
\keyword{ package }
